move the implementation to the _Mixin class
