+++ gquant/dataframe_flow/_node_flow.py
@@ -36,6 +36,34 @@ _copys = {}  # dictionary of object copy functions
 _cleanup = {}  # dictionary of clean up functions
 
 
+def dynamic_ports(ports_setup):
+
+    def calculated_ports_setup(self):
+        # note, currently can only handle one dynamic port per node
+        port_type = PortsSpecSchema.port_type
+        ports = ports_setup(self)
+        inports = ports.inports
+        dy = PortsSpecSchema.dynamic
+        for key in inports:
+            if dy in inports[key] and inports[key][dy]:
+                types = inports[key][port_type]
+                break
+        else:
+            return ports
+        if hasattr(self, 'inputs'):
+            for inp in self.inputs:
+                to_port = inp['to_port']
+                if to_port in inports and (not inports[to_port].get(dy,
+                                                                    False)):
+                    # skip connected non dynamic ports
+                    continue
+                else:
+                    inports[inp['from_node'].uid+'@'+inp['from_port']] = {
+                        port_type: types, dy: True}
+        return ports
+    return calculated_ports_setup
+
+
 def _get_nodetype(node):
     '''Identify the implementation node class. A node might be mixed in with
     other classes. Ideally get the primary implementation class.
@@ -118,6 +146,32 @@ class NodeTaskGraphMixin(object):
         #     user configurable i.e. "df" is just some python object which is
         #     typically a data container.
         self.clear_input = True
+
+    def _calculate_dynamic_input_meta(self):
+        """
+        This method will return the dynamic input metadata.
+        It will fixed the port names that is changed dynamically
+        """
+        output = {}
+        ports = self.calculated_ports_setup()
+        inports = ports.inports
+        iports_connected = self.get_connected_inports()
+        iports_cols = self.get_input_meta()
+        dy = PortsSpecSchema.dynamic
+        if hasattr(self, 'inputs'):
+            for inp in self.inputs:
+                oport = inp['to_port']
+                iport = inp['from_port']
+                out_port_name = inp['from_node'].uid+'@'+iport
+                if out_port_name in inports and inports[
+                        out_port_name].get(dy, False):
+                    if oport in iports_connected and oport in iports_cols:
+                        output[out_port_name] = copy.deepcopy(
+                            iports_cols[oport])
+                else:
+                    if oport in iports_connected and oport in iports_cols:
+                        output[oport] = copy.deepcopy(iports_cols[oport])
+        return output
 
     def __valide(self, node_output: dict):
         output_meta = self.meta_setup().outports
@@ -539,3 +593,167 @@ class NodeTaskGraphMixin(object):
             self.save_cache(output_df)
 
         return output_df
+
+    def _validate_connected_ports(self):
+        """
+        Validate the connected port types match
+        """
+        self_nodetype = _get_nodetype(self)
+        nodetype_names = [inodet.__name__ for inodet in self_nodetype]
+        if 'OutputCollector' in nodetype_names:
+            # Don't validate for OutputCollector
+            return
+
+        msgfmt = '"{task}":"{nodetype}" {inout} port "{ioport}" {inout} port '\
+            'type(s) "{ioport_types}"'
+
+        iports_connected = self.get_connected_inports()
+        iports_spec = self._get_input_ports(full_port_spec=True)
+        for iport in iports_connected.keys():
+            iport_spec = iports_spec[iport]
+            iport_type = iport_spec[PortsSpecSchema.port_type]
+            iport_types = [iport_type] \
+                if not isinstance(iport_type, Iterable) else iport_type
+
+            for ient in self.inputs:
+                # find input node edge entry with to_port, from_port, from_node
+                if not iport == ient['to_port']:
+                    continue
+                ientnode = ient
+                break
+            else:
+                intask = self._task_obj[TaskSpecSchema.inputs][iport]
+                # this should never happen
+                raise LookupError(
+                    'Task "{}" not connected to "{}.{}". Add task spec to '
+                    'TaskGraph.'.format(intask, self.uid, iport))
+
+            from_node = ientnode['from_node']
+            oport = ientnode['from_port']
+            oports_spec = from_node._get_output_ports(full_port_spec=True)
+            oport_spec = oports_spec[oport]
+            oport_type = oport_spec[PortsSpecSchema.port_type]
+            oport_types = [oport_type] \
+                if not isinstance(oport_type, Iterable) else oport_type
+
+            for optype in oport_types:
+                if issubclass(optype, tuple(iport_types)):
+                    break
+            else:
+                # Port types do not match
+                msgi = msgfmt.format(
+                    task=self.uid,
+                    nodetype=self_nodetype,
+                    inout='input',
+                    ioport=iport,
+                    ioport_types=iport_types)
+
+                msgo = msgfmt.format(
+                    task=from_node.uid,
+                    nodetype=_get_nodetype(from_node),
+                    inout='output',
+                    ioport=oport,
+                    ioport_types=oport_types)
+
+                errmsg = 'Port Types Validation\n{}\n{}\n'\
+                    'Connected nodes do not have matching port types. Fix '\
+                    'port types.'.format(msgo, msgi)
+
+                raise TypeError(errmsg)
+
+    def _validate_connected_metadata(self):
+        """
+        Validate the connected metadata match the requirements.
+        metadata.inports specify the required metadata.
+        """
+        metadata = self.meta_setup()
+
+        # as current behavior of matching in the validate_required
+        def validate_required(iport, kcol, kval, ientnode, icols):
+            node = ientnode['from_node']
+            oport = ientnode['from_port']
+            src_task = '{}.{}'.format(node.uid, oport)
+            src_type = _get_nodetype(node)
+            # incoming "task.port":"Node-type":{{column:column-type}}
+            msgi = \
+                '"{task}":"{nodetype}" produces metadata {colinfo}'.format(
+                    task=src_task,
+                    nodetype=src_type,
+                    colinfo=icols)
+
+            dst_task = '{}.{}'.format(self.uid, iport)
+            dst_type = _get_nodetype(self)
+            # expecting "task.port":"Node-type":{{column:column-type}}
+            msge = \
+                '"{task}":"{nodetype}" requires metadata {colinfo}'.format(
+                    task=dst_task,
+                    nodetype=dst_type,
+                    colinfo={kcol: kval})
+
+            header = \
+                'Meta Validation\n'\
+                'Format "task.port":"Node-type":{{column:column-type}}'
+            info_msg = '{}\n{}\n{}'.format(header, msgi, msge)
+
+            if kcol not in icols:
+                err_msg = \
+                    'Task "{}" missing required column "{}" '\
+                    'from "{}".'.format(self.uid, kcol, src_task)
+                out_err = '{}\n{}'.format(info_msg, err_msg)
+                raise LookupError(out_err)
+
+            ival = icols[kcol]
+            if kval != ival:
+                # special case for 'date'
+                if (kval == 'date' and ival
+                        in ('datetime64[ms]', 'date', 'datetime64[ns]')):
+                    return
+                else:
+                    err_msg = 'Task "{}" column "{}" expected type "{}" got '\
+                        'type "{}" instead.'.format(self.uid, kcol, kval, ival)
+                    out_err = '{}\n{}'.format(info_msg, err_msg)
+                    raise LookupError(out_err)
+
+        inputs_meta = self.get_input_meta()
+        required = metadata.inports
+
+        if not required:
+            return
+        inports = self._get_input_ports(full_port_spec=True)
+        for iport in inports:
+            if iport not in required:
+                continue
+            required_iport = required[iport]
+
+            if iport not in inputs_meta:
+                # if iport is dynamic, skip warning
+                dy = PortsSpecSchema.dynamic
+                if inports[iport].get(dy, False):
+                    continue
+                # Is it possible that iport not connected? If so iport should
+                # not be in required. Should raise an exception here.
+                warn_msg = \
+                    'Task "{}" Node Type "{}" missing required port "{}" in '\
+                    'incoming meta data. Should the port be connected?'.format(
+                        self.uid, _get_nodetype(self), iport)
+                warnings.warn(warn_msg)
+                continue
+            incoming_meta = inputs_meta[iport]
+
+            for ient in self.inputs:
+                # find input node edge entry with to_port, from_port, from_node
+                if not iport == ient['to_port']:
+                    continue
+                ientnode = ient
+                break
+            else:
+                intask = self._task_obj[TaskSpecSchema.inputs][iport]
+                # this should never happen
+                raise LookupError(
+                    'Task "{}" not connected to "{}.{}". Add task spec to '
+                    'TaskGraph.'.format(intask, self.uid, iport))
+
+            for key, val in required_iport.items():
+                validate_required(iport, key, val,
+                                  ientnode, incoming_meta)
+
